fix: move decompressBlob out of the "no-CI" build tag
